requirements.txt
asgiref==3.7.2
certifi==2023.11.17
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-wordpress-api==0.2.0
djangorestframework==3.14.0
idna==2.8
iso8601==0.1.12
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.21.0
six==1.12.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.24.3
